pyproject.toml
[project]
name = "kafka-python-demo"
version = "0.1.0"
description = "kafka-python-demo"
authors = [
    {name = "jacob", email = "jacob8753@gmail.com"},
]
dependencies = ["kafka-python>=2.1.5", "tqdm>=4.67.1", "typer>=0.15.2"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
